Add Yarn Lockfile.
Since npm's lockfile was emitting different hashes for the same
install configurations and cluttering commits, I propose using the
Yarn module install system, which has a lockfile that (hopefully)
won't do similar.

In addition, Yarn caches packages to be reinstalled quickly,
and can be used in much the same way as npm can.