Combine install scripts into single script file
The `getmeX.sh` scripts have all been replaced with a single script,
`install.sh`, that downloads and builds all uncommon required software
to the `ext` folder.

Since everything is installed in `ext`, it's necessary to prepend
`ext/bin` to the `PATH` to run commands. The alternative is to change
the target folder in `install.sh`, or install the required dependencies
manually.

The Postgres database data is now stored in a folder `db` in the root of
the project, instead of a folder based on the user's home.